This change creates an API around the gpd codebase so that users/developers can access its functionality in a variety of ways.
I introduce the gpd class which initializes the model, and provides the functionality for predicting 3-component Obspy streams. A parameters class was created to encapsulate the various options that affect the prediction function (this keeps the predict_stream method's argument list small). The helpful sliding window function was moved to a utility file so that it can be used by others.

The gpd_predict.py file was modified to use the new gpd and gpd_predict_params objects. This serves as an example of the api and ensures there is no danger of code-drift having the code in two places.

I tested the refactored code against the original code and verified the output was the same.